backend/app/agents/supervisor.py
import google.generativeai as genai
import os
from typing import Dict, Any
from dotenv import load_dotenv

load_dotenv()

class SupervisorAgent:
    def __init__(self):
        genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
        self.model = genai.GenerativeModel('gemini-1.5-pro')
        self.system_prompt = """
        You are the Supervisor Agent for SmartHire Platform (2026).
        Your role is to orchestrate recruitment workflows by delegating to sub-agents:
        - Sourcing Agent (Resume parsing, vector matching)
        - Evaluation Agent (AI interviews, scoring)
        - Data Integrity Agent (Blockchain verification)
        
        Maintain strict audit trails and ensure RAPS framework compliance.
        """

    async def process_request(self, intent: str, context: Dict[str, Any]) -> Dict[str, Any]:
        # Logic to route to sub-agents based on intent
        # For now, simulate routing logic
        prompt = f"{self.system_prompt}\nUser intent: {intent}\nContext: {context}\nProvide orchestration plan."
        response = self.model.generate_content(prompt)
        return {"action": "orchestration_plan", "details": response.text}

supervisor_agent = SupervisorAgent()
