backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pgvector
python-jose[cryptography]
passlib[bcrypt]
httpx
python-multipart
python-dotenv
google-generativeai
pytest
pytest-asyncio
